+++ src/AtomUI/TokenSystem/SizeMapDesignToken.cs
@@ -6,48 +6,48 @@ public class SizeMapDesignToken : AbstractDesignToken
    /// <summary>
    /// XXL
    /// </summary>
-   public int SizeXXL { get; set; } = 48;
+   public double SizeXXL { get; set; } = 48;
 
    /// <summary>
    /// XL
    /// </summary>
-   public int SizeXL { get; set; } = 32;
+   public double SizeXL { get; set; } = 32;
 
    /// <summary>
    /// LG
    /// </summary>
-   public int SizeLG { get; set; } = 24;
+   public double SizeLG { get; set; } = 24;
 
    /// <summary>
    /// MD
    /// </summary>
-   public int SizeMD { get; set; } = 20;
+   public double SizeMD { get; set; } = 20;
 
    /// <summary>
    ///  Same as size by default, but could be larger in compact mode
    /// </summary>
-   public int SizeMS { get; set; }
+   public double SizeMS { get; set; }
 
    /// <summary>
    /// 默认
    /// 默认尺寸
    /// </summary>
-   public int Size { get; set; } = 16;
+   public double Size { get; set; } = 16;
 
    /// <summary>
    /// SM
    /// </summary>
-   public int SizeSM { get; set; } = 12;
+   public double SizeSM { get; set; } = 12;
 
    /// <summary>
    /// XS
    /// </summary>
-   public int SizeXS { get; set; } = 8;
+   public double SizeXS { get; set; } = 8;
 
    /// <summary>
    /// XXS
    /// </summary>
-   public int SizeXXS { get; set; } = 4;
+   public double SizeXXS { get; set; } = 4;
 }
 
 [GlobalDesignToken]
@@ -56,15 +56,15 @@ public class HeightMapDesignToken : AbstractDesignToken
    /// <summary>
    /// 更小的组件高度
    /// </summary>
-   public int ControlHeightXS { get; set; }
+   public double ControlHeightXS { get; set; }
    
    /// <summary>
    /// 较小的组件高度
    /// </summary>
-   public int ControlHeightSM { get; set; }
+   public double ControlHeightSM { get; set; }
    
    /// <summary>
    /// 较高的组件高度
    /// </summary>
-   public int ControlHeightLG { get; set; }
+   public double ControlHeightLG { get; set; }
 }
